+++ Twilio/Rest/Wireless.php
@@ -17,6 +17,7 @@ use Twilio\Rest\Wireless\V1;
  * @property \Twilio\Rest\Wireless\V1 $v1
  * @property \Twilio\Rest\Wireless\V1\CommandList $commands
  * @property \Twilio\Rest\Wireless\V1\RatePlanList $ratePlans
+ * @property \Twilio\Rest\Wireless\V1\UsageRecordList $usageRecords
  * @property \Twilio\Rest\Wireless\V1\SimList $sims
  * @method \Twilio\Rest\Wireless\V1\CommandContext commands(string $sid)
  * @method \Twilio\Rest\Wireless\V1\RatePlanContext ratePlans(string $sid)
@@ -110,6 +111,13 @@ class Wireless extends Domain {
      */
     protected function contextRatePlans($sid) {
         return $this->v1->ratePlans($sid);
+    }
+
+    /**
+     * @return \Twilio\Rest\Wireless\V1\UsageRecordList
+     */
+    protected function getUsageRecords() {
+        return $this->v1->usageRecords;
     }
 
     /**