+++ Twilio/Tests/Integration/Api/V2010/Account/ApplicationTest.php
@@ -21,17 +21,13 @@ class ApplicationTest extends HolodeckTestCase {
 
         try {
             $this->twilio->api->v2010->accounts("ACXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX")
-                                     ->applications->create("friendly_name");
+                                     ->applications->create();
         } catch (DeserializeException $e) {}
           catch (TwilioException $e) {}
 
-        $values = array('FriendlyName' => "friendly_name", );
-
         $this->assertRequest(new Request(
             'post',
-            'https://api.twilio.com/2010-04-01/Accounts/ACXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX/Applications.json',
-            null,
-            $values
+            'https://api.twilio.com/2010-04-01/Accounts/ACXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX/Applications.json'
         ));
     }
 
@@ -65,7 +61,7 @@ class ApplicationTest extends HolodeckTestCase {
         ));
 
         $actual = $this->twilio->api->v2010->accounts("ACXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX")
-                                           ->applications->create("friendly_name");
+                                           ->applications->create();
 
         $this->assertNotNull($actual);
     }