+++ Twilio/Tests/Integration/Api/V2010/Account/Usage/Record/ThisMonthTest.php
@@ -53,6 +53,7 @@ class ThisMonthTest extends HolodeckTestCase {
                     {
                         "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                         "api_version": "2010-04-01",
+                        "as_of": "2019-06-24T22:32:49+00:00",
                         "category": "sms-inbound-shortcode",
                         "count": "0",
                         "count_unit": "messages",