+++ Twilio/Tests/Integration/Api/V2010/Account/Usage/RecordTest.php
@@ -52,6 +52,7 @@ class RecordTest extends HolodeckTestCase {
                     {
                         "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                         "api_version": "2010-04-01",
+                        "as_of": "2019-06-24T22:32:49+00:00",
                         "category": "totalprice",
                         "count": null,
                         "count_unit": "",