+++ Twilio/Tests/Integration/Preview/Accsecurity/Service/VerificationCheckTest.php
@@ -43,7 +43,7 @@ class VerificationCheckTest extends HolodeckTestCase {
                 "sid": "VEaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "service_sid": "VAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                "to": "+14159373912",
+                "to": "+15017122661",
                 "channel": "sms",
                 "status": "approved",
                 "valid": false,