+++ Twilio/Tests/Integration/Video/V1/Room/Participant/SubscribeRulesTest.php
@@ -41,7 +41,16 @@ class SubscribeRulesTest extends HolodeckTestCase {
                 "room_sid": "RMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "date_created": null,
                 "date_updated": null,
-                "rules": []
+                "rules": [
+                    {
+                        "type": "include",
+                        "all": true,
+                        "publisher": null,
+                        "track": null,
+                        "kind": null,
+                        "priority": null
+                    }
+                ]
             }
             '
         ));
@@ -81,7 +90,11 @@ class SubscribeRulesTest extends HolodeckTestCase {
                 "rules": [
                     {
                         "type": "exclude",
-                        "all": true
+                        "all": true,
+                        "publisher": null,
+                        "track": null,
+                        "kind": null,
+                        "priority": null
                     }
                 ]
             }